LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/autoplay-disabled-by-feature-policy.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/common/media.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/feature-policy/resources/autoplay.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/media/A4.mp4 from asking for credentials because it is a cross-origin request.


PASS Feature-Policy header: autoplay "none" has no effect on the top level document.
FAIL Feature-Policy header: autoplay "none" disallows same-origin iframes. assert_false: autoplay expected false got true
FAIL Feature-Policy header: autoplay "none" disallows cross-origin iframes. assert_false: autoplay expected false got true
